examples/distributed-key-value-store/benchmark/requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
